build: upgrade security-crypto to 1.0.0
We received a report of a crash caused by the following exception:
java.security.GeneralSecurityException: decryption failed
 	at
com.google.crypto.tink.daead.DeterministicAeadWrapper$WrappedDeterministicAead.decryptDeterministically(DeterministicAeadWrapper.java:90)
 	at
androidx.security.crypto.EncryptedSharedPreferences.decryptKey(EncryptedSharedPreferences.java:562)

We found out multiple report of the same issue in google issuetracker
for the security-crypto library, that mention that the issue should be
fixed / heavily mitigated by the use of version ≥ 1.4.0 of the underlying library tink.
https://issuetracker.google.com/issues/158234058
https://issuetracker.google.com/issues/160340797#comment11
https://github.com/google/tink/issues/321#issuecomment-641632441
As it turns out security-crypto 1.0.0 uses tink 1.5.0:
https://developer.android.com/jetpack/androidx/releases/security#1.0.0
so we upgrade to that version.

Signed-off-by: Diego Rondini <diego.rondini@kynetics.com>